Création d'un nouveau projet versionné"

+++  Tp1_stats_laure/src/tp1_stats_laure/Tp1_stats_laure.java
@@ -17,12 +17,12 @@ public class Tp1_stats_laure {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        int [] Tab =new int[6];
+        Double [] Tab =new Double[6];
         Scanner sc = new Scanner(System.in);
         Random generateurAleat = new Random(); // Initialiser le générateur aléatoire
 
         System.out.print("Saisissez une valeur (m) : ");
-        int m = sc.nextInt(); // Lire la valeur de m
+        double m = sc.nextInt(); // Lire la valeur de m
         for (int i=0;i<m;i++){
             int alea= generateurAleat.nextInt(6);
             Tab[alea]= Tab[alea]+1; // incrémenter la case correspondant dans le tableau
